requirements.txt
asgiref==3.8.1
Django==4.2.9
django-admin-interface==0.28.6
django-ckeditor==6.7.1
django-colorfield==0.11.0
django-extensions==3.2.3
django-js-asset==2.0.0
django-phonenumber-field==7.0.2
django-versatileimagefield==3.0
pillow==10.3.0
psycopg2-binary==2.9.9
python-decouple==3.7
python-magic==0.4.27
python-slugify==8.0.4
sqlparse==0.5.0
text-unidecode==1.3
typing_extensions==4.11.0
django-import-export==3.2.0
django-admin-interface
django-colorfield
easy-thumbnails
django-tinymce
django-import-export
django-crispy-forms
crispy-bootstrap5
django-tables2
django-filter
django-registration-redux
django-extensions
django_compressor
python-decouple
django==4.2.9
django-user-sessions